requirements.txt
attrs==21.4.0
cffi==1.14.4
cryptography==3.3.2
iniconfig==1.1.1
mock==4.0.3
mysql-connector-python==8.0.29
packaging==21.3
pluggy==1.0.0
protobuf==4.21.1
py==1.11.0
pycparser==2.20
PyMySQL==0.10.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.15.0
six==1.15.0
tomli==2.0.1
pygount
pytest-cov
pylint
black